Constants.loadObjectFile
